Revamped design sweep inheritance hierarchy and generator system.
The responsibility of implementing the appropriate generate_* functions
(e.g. configs, trace, etc.) is now on the sweep object. We no
longer search for the appropriately named generator module; if the
function isn't present in the sweep object, then an AttributeError
is raised.

Xenon now has its own global scope. This is to allow use commands to be
executed before a sweep has been declared, so that users can use custom
sweep types.

All sweep objects inherit from BaseDesignSweep or a subclass of it.
The default sweep type is ExhaustiveSweep. Users can subclass
BaseDesignSweep to implement their own custom sweep types. To use their
custom type, they will import their types at the top of the Xenon script
using a use command, which adds the types to Xenon global scope.

+++ tests/test_complete.py
@@ -26,8 +26,9 @@ class Common(object):
       self.assertEqual(expected, output)
 
     def tearDown(self):
-      output_dir = os.path.dirname(self.genfiles[0])
-      shutil.rmtree(output_dir)
+      if len(self.genfiles):
+        output_dir = os.path.dirname(self.genfiles[0])
+        shutil.rmtree(output_dir)
 
 class SimpleSweepParam(Common.CompleteSweepTest):
   def setUp(self):
